feat: enhance fish detection and analysis with new features and code refactoring
- Introduced new tasks in `mise.toml` for development with `--watch` functionality.
- Refactored sliders and layout in `notebook.py` for better readability.
- Split video frame processing into separate functions for clarity.
- Added support for image upload and error analysis for fish detection.
- Updated `fish.py` with contour handling enhancements.
- Added new dependencies for dev in `pyproject.toml` and updated `uv.lock` accordingly.

+++ fish.py
@@ -151,7 +151,7 @@ def draw_detections(frame, detections):
 
         # Rotated rectangle from contour analysis (cyan)
         if rotated_rect is not None:
-            box_points = cv2.boxPoints(rotated_rect).astype(int)
+            box_points = cv2.boxPoints(rotated_rect).astype(np.int32)
             cv2.drawContours(frame, [box_points], 0, (255, 255, 0), 2)
 
         # Label with both dimensions
